Fixes to install-requirements
The main bug is that packages named explicitly on the command line
without a version were installed at the latest version, instead of the
version pinned in the defaults file.

Also changed from calling pip.main to launching pip via subprocess,
which avoids issues due to pip not really being designed to run main
twice (mostly claiming that a package was not found in the verification
pass).

Since it's now safe to have install-requirements.py install pip, the
Dockerfiles are updated to use install-requirements to install
pre-requirements.