Make docker-base-gpu nvidia-docker compatible
Instead of installing a fixed version of the drivers (which need to
match the host), it sets a label to tell nvidia-docker to provide the
drivers, and sets up appropriate environment variables to allow the
mounted drivers to be located.

This should also be compatible with katsdpcontroller, which similarly
mounts GPU drivers.