Make mirror selectable
The default mirror (http://sdp-services.kat.ac.za/mirror) can be
overridden by passing `--build-arg KATSDPDOCKERBASE_MIRROR=<mirror>` on
the command line (can also be empty to disable the mirror entirely).

If the root URL of the mirror cannot be fetched, it falls back to no
mirroring. This should make it possible for anyone to build these images
even outside the SARAO offices.

Closes SR-1631.