Improvements to install-requirements.py
The main improvement is that it can properly handle requirements with a
marker but no version pin e.g. `ipaddress; python_version<'3'` and get
the pin from the default file. Other changes include

- It only considers requirements whose markers match. So for example, it
  becomes possible to specify different defaults for different Python
  versions.
- Pinned versions are added to the given specifiers rather than
  replacing them. So if for some reason requirements.txt contained
  `ipaddress >= 2` but the defaults contained `ipaddress==1.0.22` then
  it would combine them (which would fail to install) instead of
  silently removing the `>= 2` specifier.
- Names are canonicalized when matching to the defaults, so for example
  one could use `markupsafe` in the defaults but `MarkupSafe` in
  requirements.txt.
- Some PEP-8 cleanup.

Some of this is powered by the `packaging` module (which also powers
`pkg_resources`, but is vendored there), so it is added to
pre-requirements.txt to ensure that it is always available.