Make it easier to build against an alternate registry
The Dockerfiles are parametrised with the path to the registry, although
still defaulting to sdp-docker-registry. In future we could change
replace it with just a Dockerhub username for Dockerhub to build the
images. The Jenkins builder script takes the environment variable
DOCKER_REGISTRY (set in the Jenkins server configuration) and passes it
through.

As an added bonus, the build script also passes the tag being built, and
the Dockerfiles use that to specify the base image. This will fix the
problem of building branches of katsdpdockerbase itself, where
docker-base-build will be FROM docker-base-runtime:latest instead of
FROM docker-base-runtime:branch-being-built.