Add new tool for installing pinned requirements
It uses internals of pip-tools to determine recursive requirements, so
they do not need to be explicitly listed as long as they have their
versions pinned by constraints or defaults files.

Compatibility with pip is improved by supporting `-r` and `-c` inside
files, and `-c` on the command line.

pre-requirements are bumped to newer versions (to work with pip-tools),
and pip-tools is added so that install_pinned will work.