Fix for NVIDIA 460.xx drivers
This was broken by #71, which upgraded CUDA to 11.3.1 and also set
`NVIDIA_REQUIRE_CUDA>=11.3` in the runtime image. The meaning of this
setting is a bit murky, but I believe that each driver release reports a
version of CUDA is supports, and this is compared before starting the
container. The 460.39 drivers deployed on site advertise support for
CUDA 11.2, but according to the CUDA release notes, CUDA 11.0 to 11.4
all have the same requirement: 450.80.02 or later.

Since all CUDA 11 versions have the same minimum requirement, I've
changed NVIDIA_REQUIRE_CUDA to 11.0 (which matches what was in the build
image).