Bump some vulnerable versions in base-requirements
This should subsume some of the PRs that Dependabot opened. I left alone
pyjwt, numpy and dask, since they're high-risk updates and I don't want
them to be my problem.

markupsafe wasn't directly flagged by dependabot, but it flagged a
dependent in a downstream repo that needs a newer markupsafe to be
upgraded.