SPR1-2766 Support katsdptelstate 0.13
This bumps redis-py and friends to the latest versions. These should
all be accessed via katsdptelstate, which makes it safe to upgrade.

Katsdptelstate 0.13 now depends on redis-py >= 4.2, which has aioredis
built in and therefore does not require the standalone aioredis package
anymore. I've updated hiredis as well, as the redis-py 4.5.0 changelog
suggests hiredis >= 2.2.1 for improved write access.