SPR1-2766 Add lupa==1.14.1 to docker base
This is still some minor fallout of the katsdptelstate 0.13 release.

As discussed in ska-sa/katsdpimager#124, it is useful to pull lupa
into docker base as it is a dependency of `fakeredis[lua]`, and the
two packages typically go hand-in-hand in our packages.

Currently the two are used in the tests of `katsdptelstate` and
`katsdpimager`. This change avoids updating lupa in *two* places in
future (the minimum to make this beneficial :-) ).